types for blog-post
